refactor: implement deep cognitive analysis prompts
Complete redesign of AI feedback prompt system based on Deep Cognitive Analysis principles.
Combines v2 compression improvements with v3 clinical analysis framework.

**Core Improvements**:
1. **Cognitive Mechanism Breakdown**
   - Analyze situation → interpretation → behavior → outcome causal chain
   - Explore how assumptions, fears, biases, and mental models shape interpretations

2. **Leverage Points**
   - Identify 1-2 small changes that create maximum behavioral change
   - Focus on high-leverage intervention points vs. entire system

3. **Tone Shift**: Clinical but supportive
   - Cognitive systems analyst (not coach)
   - Analytical, sharp, specific
   - Mechanism explanation (not motivation/encouragement)

4. **Enhanced Prohibitions**:
   - Structurally block generic self-improvement advice
   - Ban "try harder", "be consistent" platitudes
   - Forbid overly positive coaching tone
   - Prevent productivity framework rehashing

5. **Strengthened Output Format**:
   - To-do: Ban "try this", require why/what/when
   - Insights: "You said [actual quote]. This reveals [mechanism]"
   - Patterns: Analyze thought habits, triggers, emotional cycles, decision loops
   - Improvements: Require why/what/how/measurement
   - Suggestions: Sharp questions (expose blind spots, trigger self-awareness)

6. **Token Optimization**:
   - 50% token reduction (removed system + user redundancies)
   - Resolved contradictions ("must have 5 sections" vs "omit if no data")
   - Explicit leveling (founder/maker, high self-reflection)
   - Enforced "quote actual expressions 1+ times"
   - Removed "recent 3-7 days data" requirement → flexible content-based pattern analysis

**Changed Files**:
- src-tauri/src/llm/prompts.rs: Deep cognitive analysis system
- src/constants/aiPrompts.ts: Deep cognitive analysis prompts

**Expected Impact**:
- Superficial advice → Deep cognitive mechanism-based analysis
- Generic suggestions → High-leverage intervention-focused execution plans
- Gentle coaching → Sharp, specific insights

+++ src-tauri/src/llm/engine.rs
@@ -38,7 +38,7 @@ impl LlamaCppEngine {
         };
         // Auto-detect reasonable CPU threads
         let threads = num_cpus::get_physical().max(1);
-        s.config.cpu_threads = threads as usize;
+        s.config.cpu_threads = threads;
         Ok(s)
     }
